minor fix and rename quat_linear  folders
