+++ .github/workflows/python-package.yml
@@ -63,7 +63,7 @@ jobs:
         # stop the build if there are Python syntax errors or undefined names
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 ./qllm/quant --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        flake8 ./qllm/qllm/modeling/q_layers --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
         flake8 ./qllm/quantization --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
         flake8 ./qllm/utils --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: build pypi package
