+++ qllm/modeling/q_layers/quant_linear.py
@@ -74,7 +74,7 @@ class DequantAndUnpack(torch.autograd.Function):
 
     @staticmethod
     def forward(ctx, qweight, scales, qzeros, groupsize, bits, in_features, g_idx, act_order):
-        compatible_with_autogptq = int(os.environ.get('compatible_with_autogptq', "0"))
+        compatible_with_autogptq = int(os.environ.get("compatible_with_autogptq", "0"))
         if has_module_XbitOps and qweight.is_cuda and not act_order:
             return XbitOps.dequant(qweight, scales, qzeros, groupsize, bits, in_features, compatible_with_autogptq)
         scales = scales.reshape(-1, 1, scales.shape[-1])
@@ -117,7 +117,7 @@ class DequantAndUnpack(torch.autograd.Function):
 
 
 def QuantLinearTorchFunction_forward(input, qweight, scales, qzeros, g_idx, bits, groupsize, in_features, act_order):
-    compatible_with_autogptq = int(os.environ.get('compatible_with_autogptq', "0"))
+    compatible_with_autogptq = int(os.environ.get("compatible_with_autogptq", "0"))
     if not act_order and not torch.onnx.is_in_onnx_export() and input.reshape(-1, input.shape[-1]).shape[0] <= 4 and bits == 4 and groupsize==128:
         return XbitOps.gemv(input, qweight, scales, qzeros, groupsize, bits, in_features, compatible_with_autogptq)
     weight = DequantAndUnpack().apply(qweight, scales, qzeros, groupsize, bits, in_features, g_idx, act_order)
